Refactor excel download into separate classes
