+++ app/static/src/app/store/run/actions.ts
@@ -5,7 +5,7 @@ import { AppState, AppType } from "../appState/state";
 import userMessages from "../../userMessages";
 import type { OdinRunDiscreteInputs, OdinRunResultDiscrete, OdinRunResultOde } from "../../types/wrapperTypes";
 import { Odin, OdinRunnerOde, OdinUserType } from "../../types/responseTypes";
-import { WodinExcelDownload } from "../../wodinExcelDownload";
+import { WodinModelOutputDownload } from "../../excel/wodinModelOutputDownload";
 import { ModelFitAction } from "../modelFit/actions";
 import { RunGetter } from "./getters";
 import { SensitivityMutation } from "../sensitivity/mutations";
@@ -136,8 +136,8 @@ export const actions: ActionTree<RunState, AppState> = {
         const { commit } = context;
         commit(RunMutation.SetDownloading, true);
         setTimeout(() => {
-            new WodinExcelDownload(context, payload.fileName, payload.points)
-                .downloadModelOutput();
+            new WodinModelOutputDownload(context, payload.fileName, payload.points)
+                .download();
             commit(RunMutation.SetDownloading, false);
         }, 5);
     },
